Add IndexDescriptor class to public interface
